20230320 16:36

+++ Spring-LmsProject/src/main/java/com/java/project/repo/VideoRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.java.project.vo.Video;
 
 @Repository
-public interface VideoRepository extends JpaRepository<Video, Long> {
+public interface VideoRepository extends JpaRepository<Video, Integer> {
     
     @Query("SELECT v.title, v.fname, v.duration, v.description, q.leveltest FROM Video v INNER JOIN Quiz q ON v.lvl_code=q.lvl_code WHERE v.lvl_code=:lvlCode")
     public Map<String, Object> getVideoByLvl_code(@Param("lvlCode") int lvl_code);
